indosbert/requirements.txt
faiss
matplotlib
numpy
pandas
python-dotenv
seaborn
sentence-transformers
streamlit
supabase
torch